requirements.txt
Django>=5.0.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
pycryptodome>=3.20.0
python-dotenv>=1.0.0
Pillow>=10.2.0
django-filter>=23.5
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.9
cryptography>=42.0.2
django-allauth>=0.61.1
gunicorn>=21.2.0
whitenoise>=6.6.0 